Moved instagram class to utils file
